Add files via upload
The Purpose of this site is to demonstrate my ability to create a functional and efficient mobile site using jQuery mobile libraries and methods to modify html files for mobile use. This site is a passion project about video games, and I enjoyed every minute creating something I love. I believe, it offers a unique set of short reviews of popular video games of 2023.